+++ src/errors/body.ts
@@ -3,7 +3,7 @@ import { ErrorWithStatusCode } from "./status";
 
 export class ErrorWithBody extends ErrorWithStatusCode {
   public body?: IBody;
-  constructor(statusCode: number, body?: IBody) {
+  public constructor(statusCode: number, body?: IBody) {
     super(statusCode, body ? body.toString() : undefined);
     this.body = body;
   }