+++ src/errors/status.ts
@@ -1,7 +1,7 @@
 export class ErrorWithStatusCode extends Error {
   public statusCode: number;
 
-  constructor(statusCode: number, message?: string) {
+  public constructor(statusCode: number, message?: string) {
     super(message);
     this.statusCode = statusCode;